feat: added command up
