+++ src/handlers-command/up/up.js
@@ -1,0 +1,6 @@
+import { join } from 'path';
+import { workDirectory } from '../../helpers/helpers.js';
+
+export const handleUp = () => {
+  workDirectory.setCWD(join(workDirectory.getCWD(), '..'));
+};
