--- src/helpers/print-output.js
@@ -1,3 +1,0 @@
-import { stdout } from 'process';
-
-export const printOutput = (message) => stdout.write(message + '\n');
