+++ src/parse-command/parse-command.js
@@ -1,15 +1,17 @@
-import { exitApp } from '../bye-user/bye-user.js';
-import { printOutput } from '../helpers/print-output.js';
-
-export const printInvalidCommand = () => printOutput('Invalid input');
+import { exitApp } from '../handlers-command/exit/exit.js';
+import { handleUp } from '../handlers-command/up/up.js';
+import { printOutput } from '../helpers/helpers.js';
 
 export const parseCommand = (command) => {
   switch (command) {
     case '.exit':
       exitApp();
       return;
+    case 'up':
+      handleUp();
+      break;
 
     default:
-      printInvalidCommand();
+      printOutput('Invalid input');
   }
 };
